requirements.txt
amqp                   2.5.1
backports.csv          1.0.7
billiard               3.6.1.0
celery                 4.3.0
certifi                2019.6.16
chardet                3.0.4
defusedxml             0.6.0
diff-match-patch       20181111
Django                 1.11
django-crispy-forms    1.7.2
django-formtools       2.1
django-haystack        2.8.1
django-import-export   1.2.0
django-ranged-response 0.2.0
django-redis           4.10.0
django-simple-captcha  0.5.12
django-tinymce         2.8.0
dnspython              1.16.0
et-xmlfile             1.0.1
eventlet               0.25.1
fdfs-client-py         1.2.6
future                 0.15.2
greenlet               0.4.15
httplib2               0.13.1
idna                   2.8
importlib-metadata     0.21
itsdangerous           1.1.0
jdcal                  1.4.1
jieba                  0.39
kombu                  4.6.4
monotonic              1.5
more-itertools         7.2.0
mutagen                1.42.0
odfpy                  1.4.0
openpyxl               2.6.3
Pillow                 6.1.0
pip                    19.2.3
pycryptodomex          3.7.2
PyMySQL                0.9.3
python-alipay-sdk      1.10.1
pytz                   2019.2
PyYAML                 5.1.2
redis                  3.3.8
requests               2.22.0
setuptools             41.2.0
six                    1.12.0
SQLAlchemy             1.1.15
sqlparse               0.3.0
tablib                 0.13.0
urllib3                1.25.3
vine                   1.3.0
wheel                  0.33.6
Whoosh                 2.7.4
xlrd                   1.2.0
xlwt                   1.3.0
zipp                   0.6.0